+++ lib/utils/constants/app_assets.dart
@@ -7,7 +7,7 @@ class AppAssets {
   static const String iconWhite = "assets/appIcon_white.png";
   static const String isologoAppWhite = "assets/logos/isologo-app-white.png";
   static const String logo = "assets/logos/logo-app.png";
-  static const String logoCompanyWhite = "assets/logos/logo-company-white.png";
+  static const String logoCompanyWhite = "assets/logos/logo_company_white.png";
   static const String logoSloganSpanish = "assets/logos/logo-slogan-spanish.png";
   static const String logoSloganEnglish = "assets/logos/logo-slogan-english.png";
   static const String facebookLogo = 'assets/logos/third-party/facebook.jpg';